finish get uniform sample polars

+++ ma_sh/Cpp/src/idx.cpp
@@ -5,7 +5,7 @@ const torch::Tensor toBoundIdxs(const torch::Tensor &data_counts) {
                                  .toType(data_counts.scalar_type())
                                  .to(data_counts.device());
 
-  for (size_t i = 1; i < bound_idxs.sizes()[0]; ++i) {
+  for (int i = 1; i < bound_idxs.sizes()[0]; ++i) {
     bound_idxs[i] = data_counts[i - 1] + bound_idxs[i - 1];
   }
   return bound_idxs;
